requirements.txt
torch 
tensorboardX 
tensorboard 
gym 
PyYaml 
pettingzoo 
# Box2D 
box2d-kengz # for compute canada only 
pandas # used in communication analysis 
matplotlib # used in communication analysis 
scikit-learn # used in communication analysis 
npy-append-array 
pygame
protobuf==3.20.1